Create Docker Container Publish Workflow (#2)
This change adds a GitHub Actions Workflow intended to package the code into a Docker container image and publish it to the GitHub Container Registry. Details on the workflow are included with links in comments in the workflow file.

Note that this workflow should not trigger until a release is created.